+++ docs/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+4.1.0
+-----
+
+Features:
+  * Add Binance interchain support
+Development:
+  * Add integration tests for Binance feature
+
 4.0.1
 -----